requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
httpx==0.27.2
groq==0.13.0
email-validator>=2.1.0
PyJWT==2.9.0
bcrypt==4.1.2
sqlmodel==0.0.21
alembic==1.13.2
click==8.1.7
